refactor: ranking 관련 파일명 수정 #65
