+++ src/components/ranking/RankingListInfo.tsx
